+++ src/com/cloudogu/gitopsbuildlib/deployment/helm/helmrelease/ArgoCDRelease.groovy
@@ -16,34 +16,40 @@ class ArgoCDRelease extends HelmRelease {
         Map helmConfig = gitopsConfig.deployments.helm
         String application = gitopsConfig.application
 
-        String helmRelease = ""
         if (helmConfig.repoType == 'GIT') {
-            helmRelease = createResourcesFromGitRepo(gitopsConfig, application, namespace, mergedValuesFile)
+            return createResourcesFromGitRepo(gitopsConfig, application, namespace, mergedValuesFile)
         } else if (helmConfig.repoType == 'HELM') {
-            helmRelease = createResourcesFromHelmRepo(gitopsConfig, application, namespace, mergedValuesFile)
+            return createResourcesFromHelmRepo(gitopsConfig, application, namespace, mergedValuesFile)
+        } else if (helmConfig.repoType == 'LOCAL') {
+            return createResourcesFromLocalRepo(gitopsConfig, application, namespace, mergedValuesFile)
         }
-        return helmRelease
+        return null // Validated in base class Helm
     }
 
     private String createResourcesFromGitRepo(Map gitopsConfig, String application, String namespace, String mergedValuesFile) {
         Map helmConfig = gitopsConfig.deployments.helm
 
-        def chartPath = ''
+        def chartPath = "${script.env.WORKSPACE}/.helmChartTempDir/chart/"
         if (helmConfig.containsKey('chartPath')) {
-            chartPath = helmConfig.chartPath
+            chartPath += helmConfig.chartPath
         }
 
         return createHelmRelease(gitopsConfig, chartPath as String, application, namespace, gitopsConfig.buildImages.helm, mergedValuesFile)
     }
 
     private String createResourcesFromHelmRepo(Map gitopsConfig, String application, String namespace, String mergedValuesFile) {
-        return createHelmRelease(gitopsConfig, gitopsConfig.deployments.helm.chartName, application, namespace, gitopsConfig.buildImages.helm, mergedValuesFile)
+        String chartPath = "${script.env.WORKSPACE}/.helmChartTempDir/chart/${gitopsConfig.deployments.helm.chartName}"
+        return createHelmRelease(gitopsConfig, chartPath, application, namespace, gitopsConfig.buildImages.helm, mergedValuesFile)
+    }
+    
+    private String createResourcesFromLocalRepo(Map gitopsConfig, String application, String namespace, String mergedValuesFile) {
+        return createHelmRelease(gitopsConfig, gitopsConfig.deployments.helm.chartPath, application, namespace, gitopsConfig.buildImages.helm, mergedValuesFile)
     }
 
     private String createHelmRelease(Map gitopsConfig, def chartPath, String application, String namespace, def helmImageConfig, String mergedValuesFile) {
         String helmRelease = ""
         dockerWrapper.withDockerImage(helmImageConfig) {
-            String templateScript = "helm template ${application} ${script.env.WORKSPACE}/.helmChartTempDir/chart/${chartPath} -n ${namespace} --kube-version ${gitopsConfig.k8sVersion} -f ${mergedValuesFile}"
+            String templateScript = "helm template ${application} ${chartPath} -n ${namespace} --kube-version ${gitopsConfig.k8sVersion} -f ${mergedValuesFile}"
             helmRelease = script.sh returnStdout: true, script: templateScript
         }