+++ src/com/cloudogu/gitopsbuildlib/deployment/helm/repotype/LocalRepo.groovy
@@ -1,0 +1,12 @@
+package com.cloudogu.gitopsbuildlib.deployment.helm.repotype
+
+class LocalRepo extends RepoType {
+
+    LocalRepo(def script) {
+        super(script)
+    }
+
+    @Override
+    void prepareRepo(Map gitopsConfig, String helmChartTempDir, String chartRootDir) {
+    }
+}