+++ CHANGELOG.md
@@ -13,6 +13,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Specify version for `helm template` through the `k8sVersion` parameter
 
+### Fixed
+- Don't fail when no `values-shared.yaml` is provided
+
 ## [0.2.0](https://github.com/cloudogu/gitops-build-lib/releases/tag/0.2.0) - 2023-03-10
 
 ### Added